requirements.txt
fastmcp
google-generativeai